requirements.txt
# Geospatial / tile serving (inherited from parent project)
fastapi==0.135.1
uvicorn==0.34.2
titiler[standard]
rasterio
numpy
GDAL==3.8.4

# Image processing
Pillow
opencv-python-headless
scikit-image

# Deep learning
torch
torchvision

# SR models
basicsr          # Real-ESRGAN dependency
realesrgan

# Metrics
lpips            # Perceptual loss / LPIPS metric

# Data / utilities
requests
tqdm
pandas
